moved UniqueValuesLegendColorMapRow.svelte
