+++ src/components/controls/UniqueValuesLegendColorMapRow.svelte
